Refactor fix_data.yaml (#114192)
* Refactor fix_data.yaml

* ++

* lib, not lib/src

* Nit

* Nit

* Update packages/flutter/test_fixes/material/app_bar_theme.dart


* Fix copy paste errors
