+++ packages/flutter/test_fixes/widgets/build_context.dart
@@ -1,0 +1,17 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter/widgets.dart';
+
+void main() {
+  // Changes made in https://github.com/flutter/flutter/pull/44189
+  const BuildContext buildContext = Element(myWidget);
+  buildContext.inheritFromElement(ancestor);
+  buildContext.inheritFromWidgetOfExactType(targetType);
+  buildContext.ancestorInheritedElementForWidgetOfExactType(targetType);
+  buildContext.ancestorWidgetOfExactType(targetType);
+  buildContext.ancestorStateOfType(TypeMatcher<targetType>());
+  buildContext.rootAncestorStateOfType(TypeMatcher<targetType>());
+  buildContext.ancestorRenderObjectOfType(TypeMatcher<targetType>());
+}
