+++ packages/flutter/test_fixes/widgets/element.dart
@@ -1,0 +1,17 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter/widgets.dart';
+
+void main() {
+  // Changes made in https://github.com/flutter/flutter/pull/44189
+  const Element element = Element(myWidget);
+  element.inheritFromElement(ancestor);
+  element.inheritFromWidgetOfExactType(targetType);
+  element.ancestorInheritedElementForWidgetOfExactType(targetType);
+  element.ancestorWidgetOfExactType(targetType);
+  element.ancestorStateOfType(TypeMatcher<targetType>());
+  element.rootAncestorStateOfType(TypeMatcher<targetType>());
+  element.ancestorRenderObjectOfType(TypeMatcher<targetType>());
+}
